moving FileTarget archival methods to their own namespace.
