+++ src/NLog/NLog.wp71.csproj
@@ -353,6 +353,16 @@
     <Compile Include="Targets\EventLogTarget.cs" />
     <Compile Include="Targets\FileArchivePeriod.cs" />
     <Compile Include="Targets\FileTarget.cs" />
+    <Compile Include="Targets\FileTargetArchival\Archival.cs" />
+    <Compile Include="Targets\FileTargetArchival\DateAndSequenceArchival.cs" />
+    <Compile Include="Targets\FileTargetArchival\DateArchival.cs" />
+    <Compile Include="Targets\FileTargetArchival\DateArchivalCommon.cs" />
+    <Compile Include="Targets\FileTargetArchival\DynamicFileArchive.cs" />
+    <Compile Include="Targets\FileTargetArchival\FileNameTemplate.cs" />
+    <Compile Include="Targets\FileTargetArchival\IArchivalCallbacks.cs" />
+    <Compile Include="Targets\FileTargetArchival\IArchivalOptions.cs" />
+    <Compile Include="Targets\FileTargetArchival\RollingArchival.cs" />
+    <Compile Include="Targets\FileTargetArchival\SequenceArchival.cs" />
     <Compile Include="Targets\LineEndingMode.cs" />
     <Compile Include="Targets\LogReceiverWebServiceTarget.cs" />
     <Compile Include="Targets\MailTarget.cs" />
